Introduce a "stream reader" abstraction
Useful for keeping the state of the pointer instead of passing it
around from function to function, as some parts of the format are
placed at different positions depending on previous parts. Eg, if
the page count is not know, there are no four bytes after that
field to indicate the page count, so subsequent parts should verify
whether the page count is known or not and adjust the pointer passed
to them. This stream reader abstraction solves this problem quite
nicely for the moment.